Store and retrieve lotecation
- location record is now returned when db is queried for lotes
- database now stores lng/lat locations

+++ db/migrations/20170605115532_add_geocoordinate.js
@@ -1,8 +1,8 @@
 
 exports.up = function(knex, Promise) {
   return knex.schema.table('locations', (table) => {
-    table.float('latitude');
-    table.float('longitude');
+    table.double('latitude');
+    table.double('longitude');
   });
 };
